Add (array) $data to all mail events
Sometimes it is useful to have all the data that is passed to the mail view within $data available in mail events, specifically mailer.send. It also allows passing along data without having to "abuse" the $message object by setting properties on it which are not related. 

I recently had to log the output of mails that were sent along with other relevant data, that was only known within the Mail closure, so, not available within the Event listener. 
A very simple example:

```php
// Plugin.php
public function boot()
{
        // This is how the event listener signature works before this PR, whereas after the PR we would also be able to access $data
        // After PR: \Event::listen('mailer.send', function ($mailer, $view, \Illuminate\Mail\Message $message, $data)
        \Event::listen('mailer.send', function ($mailer, $view, \Illuminate\Mail\Message $message)
        {
            $log = new MailLog();
            // We do not have access to $data here, because it is not passed to the listener
            $log->transactionId = ?; 

            // With the PR
            $log->transactionId = $data['transactionId'];
        }
}

// My\Plugin\Jobs\SendInvoice.php
class SendInvoice
{
    public function fire($job, $userId)
    {
        // ...
        $transaction = $this->createTransaction($user);
        $data = ['checkoutUrl' => $transaction->checkoutUrl, 'transactionId' => $transaction->id];

        \Mail::send('my.plugin::mail.invoice', $data, function($message) 
        {
            // ....
        }
    }
}



```