Extend withDeferred() to allow pivot conditions
If a pivot has additional fields that need to be considered when fetching related data using withDeferred() there must be a way to add those conditions. The original relation query is modified by withDeferred() in such a way that it's impossible to keep any where() clauses added to it. Hence this workaround, which may not seem elegant but does the job.