Updated artisan service provider to be compatible with laravel (#60)
In newer versions of Laravel the default application commands are lazy loaded in. Because of this they've switched from using the facade accessor to the class string.

This change was introduced in this commit: https://github.com/laravel/framework/commit/82b212c1d797e73f658ef461b11796ca75707361

Credit to @jaxwilko for spotting.