Fix support for route:list & route:cache commands
Moves module route registration into the boot method where it's supposed to go.

> You should never attempt to register any event listeners, routes, or any other piece of functionality within the register method.
- https://laravel.com/docs/9.x/providers#writing-service-providers

This previous logic (where modules setup a deferred registration of their routes by binding to the App::before() event (injecting into the router immediately prior to dispatching requests; and this binding itself happens in the register() method of the serviceprovider) was originally added in order to support plugins being able to override the greedy final route provided by the CMS module. See https://github.com/wintercms/winter/commit/04bbad320fb033e2e0ebde001b58a469ff6164ec for more details.

This logic overcomplicated things however and caused issues with Laravel tools that expected to have access to all routes regardless of whether an actual request was about to be dispatched or if the application was actually being run in the CLI. In October 2.0 the approach to add another method and layer of hackiness in the form of registerLateRoutes() (which manually dispatches a blank request in order to trick the app::before into firing and registering module routes); however this solution is overcomplicated and too brittle in my personal opinion so instead I have opted to register module routes in a more Laravel-like way, where they are registered when the ServiceProvider is booted.

At some point in the future we may want to get rid of the App::before() / after() (and corresponding router events as well) altogether since there are usually better ways to achieve similar results without using them and completely non-obviously those events do not run in the CLI at all since there is no request being dispatched.