Improve Halcyon SectionParser logic
In addition to adding method signature typehints and general code tidying; this commit improves the ability of the SectionParser to handle extraneous section separators in appearing in settings, code, & markup content.

With these changes; if extra separators are found in the markup section, instead of ignoring the extra sections they will be considered part of the markup section. Additionally, if extra sections are found in the markup section and there was previously only 1 or 2 sections total (markup only or settings and markup) the SectionParser::render() method will force there to be 3 sections in order to guarantee additional separators found in the markup section stay in the markup section.

Another change to the SectionParser::render() method is that it will now throw an InvalidArgumentException if section separators are found in either the settings section or the code section as it is not currently possible to avoid accidentally parsing separators in those locations as if they were valid and intentional separators.

This is somewhat related to https://github.com/octobercms/library/commit/c393c5ce9ca2c5acc3ed6c9bb0dab5ffd61965fe although instead of just dropping all possible sections within markup and keeping the last one - this change implements more robust support for section separators appearing in places where they shouldn't.

(cherry picked from commit 97c5da24518277561d96d09c080033c03048c110)