Fire mailer.beforeRegister event in mailer() method (#79)
Related: https://github.com/wintercms/winter/pull/522

This fix is needed because if the event is fired in resolve() method, the $name variable has already been defined (from a call to getDefaultDriver() in the base mailer() method)

ref. https://github.com/laravel/framework/blob/9.x/src/Illuminate/Mail/MailManager.php#L70

Note: if the $name argument is provided to the mailer() method, the problem does not occur because the config is forced to that driver name.