Add Array Source trait (#87)
This commit introduces the ability to create an "array source" model which provides arbitrary records via an array. These models will act, for all intents and purposes, as Eloquent models, but do not require a database.

The implementation of this feature is inspired by the Sushi library by Caleb Porzio (https://github.com/calebporzio/sushi) but modified to work in Winter's architecture.

This commit also introduces a change to the testing suite, to cater for the facades used in this trait, and now creates a dummy application when running tests, which resolves a long-standing issue where facades in the test cases did not work.