Bring Translator closer to base Laravel functionality.
Refactors the "get" override to use the base "get" method for the most part, removes the "choice" method override and adds the "set" method in Translator.

Changes to the FileLoader will allow our original fix for the sublocales in the "choice" method to work even with the base Laravel functionality.

The FileLoader will now attempt to find locales in the format "xx_XX" as per the original Laravel functionality and then also attempt "xx-xx" as per the traditional usage in Winter.