Automatically purge dynamic properties added to models
When adding dynamically created properties to models, you have two choices:

1. You can either do it through direct assignment which eventually passes through Laravel's attribute system and stores the dynamic property in the $attributes property, thus flagging it for insertion into the database like a regular property; or
2. You can add it through `addDynamicProperty()` which currently ultimately ends up doing the same thing as 1 except that it also gets added to the list of dynamic properties on the model through the ExtendableTrait.

The key difference in choosing to use dynamic properties instead of model attributes usually comes down to whether or not you want your newly added property to be stored in the database. Most commonly, you would choose to use addDynamicProperty() when adding a logical property that's used for performing logic on the model (i.e. adding a ModelBehavior to the model with extendClassWith() that requires properties to be present in order to configure itself).

This assumption is present in the PurgeableBehavior itself currently (see https://github.com/wintercms/storm/blob/c972c1f50cce9c5c150278facb86e1caf9dd69ed/src/Database/Behaviors/Purgeable.php#L34-L36) which automatically adds any existing defined dynamic properties to the purgeable array during the process of booting the PurgeableBehavior when it is added to a class (a key part of the original design considerations: https://github.com/octobercms/library/pull/324#issuecomment-395419369). 

In addition to that, the Winter.Translate plugin has to go out of its way to explicitly add the Purgeable behavior whenever it adds it's own TranslateableModel behavior (see https://github.com/wintercms/wn-translate-plugin/blob/344711704f8ec0556295a02714cb2a6f48f9dd27/Plugin.php#L79-L81). 

It has also popped up in issues (https://github.com/octobercms/october/issues/3433) before as a source of developer confusion. 

Given all of the above information, it makes sense to have the Model class itself automatically add the PurgeableBehavior if not already present as soon as a single dynamic property is defined so that all dynamic properties can be automatically added to the list of purgeable properties as soon as they are defined.

In the future, it might make sense to just make the Purgeable trait itself included by default in the base model in order to not have to dynamically & automatically add the Purgeable behavior under these conditions, but I leave that as a task for later.