All scoped and local extensions.
This allows one to create either a scoped global extension through "static::extend", which will apply to all newly-constructed instances of a class going forward, or they can also use "$object->extend" to apply an extension to the single instance.

This allows access to protected and private methods within an instance, allowing full customisation of a given instance if it's allowed to be extendable.