Fix missing attributes in FormBuilder inputs
Commit https://github.com/wintercms/storm/commit/2ef32346682c768a9d457c4dba6796f34b644eab made some adjustments to the input method arguments that resulted in the some attributes not being populated if the name of the input was not provided. This commit changes the structure of the condition so that these attributes are properly filled in.

I've also redone the tests using some custom assertions so that the order of the attributes does not matter when testing. The previous tests were a bit flimsy in this regard.

Fixes https://github.com/wintercms/winter/issues/864