Local extensions called in static "extends" fire after construct
Local extensions, by design, are made against a fully constructed extendable class. However, previously, if called within the scope of a static `extends` call (which is effectively a constructor extension), it was firing immediately which meant it fired before behaviors were loaded.

This will store any callbacks for local extensions fired in a constructor extension until after the construction is completed.