Allow array-sourced models to provide records using generators
This removes the hard "array" return type for getRecords, and allows people to use generators (or other iterable types) to provide records. This should still be backwards compatible with other uses of the ArraySource trait that may already have a "getRecords" method with an array return type, due to covariance rules in PHP.

It should be noted that each record must still be an array, it's only the "container" that can be traversable.